trying to get values to show up

+++ control/include/drivers/PCA9858PWJ.hpp
@@ -1,15 +1,13 @@
 #include "I2C.hpp"
 //#define CONTROL 0xE2
 
-class PCA9858PWJ {
-    public: 
-        uint16_t readRegister(); 
-        void writeRegister(uint16_t data); 
-        PCA9858PWJ(I2C& sharedI2C);
-
-    
+class PCA9858PWJ
+{
+public:
+    uint16_t readRegister();
+    void writeRegister(uint8_t data);
+    PCA9858PWJ(I2C &sharedI2C);
 
 private:
-    I2C& _i2c;
-
+    I2C &_i2c;
 };
